Fully Function DB caching system

+++ src/db_cache.py
@@ -6,17 +6,19 @@ from datetime import timedelta
 from pickle import dumps, loads
 
 class DBCache:
-    __conn = None
+    _conn = None
+    _is_connected = False
     __cursor = None
     __cache_store = {}
     __cache_path = ''
     __last_sql = ''
     __last_params = {}
 
-    def __init__(self, cache_dir, filename):
-        self.__cache_path = join(cache_dir, filename)
-        self.__conn = sqlite3.connect(self.__cache_path)
-        self.__cursor = self.__conn.cursor()
+    def __init__(self, cache_path):
+        self.__cache_path = cache_path
+        self._conn = sqlite3.connect(self.__cache_path)
+        self._is_connected = True
+        self.__cursor = self._conn.cursor()
         self.__create_table()
         self.__prune()
     
@@ -39,7 +41,7 @@ class DBCache:
     
     def __commit(self):
         ''' Commits to the cache '''
-        self.__conn.commit()        
+        self._conn.commit()        
 
 
     def _has(self, key):
@@ -53,10 +55,11 @@ class DBCache:
         '''
             Prune away expired data from cache table
         '''
-        self.__last_sql = 'DELETE FROM cache WHERE expiration < :now'
-        self.__last_params = {'now': time()}
-        self.__execute()
-        self.__commit()
+        if self._is_connected:
+            self.__last_sql = 'DELETE FROM cache WHERE expiration < :now'
+            self.__last_params = {'now': time()}
+            self.__execute()
+            self.__commit()
     
 
     def _get(self, key):
@@ -64,12 +67,15 @@ class DBCache:
             Retrieves data (key, value) if exists in Database cache storage
             Returns: data | None
         '''
-        self.__prune()
-        self.__last_sql = 'SELECT * FROM cache where key = :key'
-        self.__last_params = {'key': key}
-        cursor = self.__execute()
-        selected_row = cursor.fetchone()
-        return loads(selected_row[1]) if selected_row else None
+        if self._is_connected:
+            self.__prune()
+            self.__last_sql = 'SELECT * FROM cache where key = :key'
+            self.__last_params = {'key': key}
+            cursor = self.__execute()
+            selected_row = cursor.fetchone()
+            return loads(selected_row[1]) if selected_row else None
+    
+        return None
 
     def _put(self, key, value, exp_mins=10):
         '''
@@ -78,7 +84,7 @@ class DBCache:
         '''
         __expiration = timedelta(minutes=exp_mins).total_seconds()
 
-        if __expiration:
+        if __expiration and self._is_connected:
             expiration = time() + __expiration
             value = dumps(value)
             self.__last_sql = '''
@@ -98,7 +104,7 @@ class DBCache:
         '''
         value = self._get(key)
 
-        if value:
+        if value and self._is_connected:
             self.__last_sql = 'DELETE FROM cache WHERE key = :key'
             self.__last_params = {'key': key}
             self.__execute()
@@ -106,12 +112,12 @@ class DBCache:
 
         return value
     
-    def _flush(self):
+    def _drop_connection(self):
         '''
-            Clears file cache storage
-            Returns: bool
+            Drops the connection to the database
         '''
-        pass
+        self._conn.close()
+        self._is_connected = False
 
     def __del__(self):
-        self.__conn.close()
+        self._conn.close()
